+++ libraries/template-haskell/Language/Haskell/TH/Quote.hs
@@ -20,7 +20,7 @@ module Language.Haskell.TH.Quote(
         dataToQa, dataToExpQ, dataToPatQ
     ) where
 
-import Language.Haskell.TH.Syntax
+import Language.Haskell.TH.Syntax hiding (readFile)
 
 -- | The 'QuasiQuoter' type, a value @q@ of this type can be used
 -- in the syntax @[q| ... string to parse ...|]@.  In fact, for